+++ AppCenter/AppCenter/include/MSAppCenter.h
@@ -1,0 +1,1 @@
+../MSAppCenter.h