+++ AppCenterAnalytics/AppCenterAnalytics/include/MSAnalytics.h
@@ -1,0 +1,1 @@
+../MSAnalytics.h