+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### App Center
 
 * **[Improvement]** Use namespaced `NSUserDefaults` keys with the **MSAppCenter** prefix for all the keys set by the SDK. Fixed a few keys missing namespace.
+* **[Feature]** Swift Package Manager support for Core and Analytics modules.
+
 ___
 
 ## Version 3.1.1